Rename navigation -> structures
