+++ src/index.tsx
@@ -61,8 +61,8 @@ import { giftCardsSectionUrlName } from "./giftCards/urls";
 import { apolloClient, saleorClient } from "./graphql/client";
 import { useLocationState } from "./hooks/useLocationState";
 import { commonMessages } from "./intl";
-import NavigationSection from "./navigation";
-import { structuresListPath } from "./navigation/urls";
+import NavigationSection from "./structures";
+import { structuresListPath } from "@dashboard/structures/urls";
 import { NotFound } from "./NotFound";
 import OrdersSection from "./orders";
 import PageSection from "./modeling";
