+++ src/structures/components/MenuDetailsPage/index.ts
