+++ src/structures/components/MenuItemDialog/options.ts
@@ -1,5 +1,5 @@
 import { sectionNames } from "@dashboard/intl";
-import { MenuItemType } from "@dashboard/navigation/components/MenuItemDialog";
+import { MenuItemType } from "@dashboard/structures/components/MenuItemDialog";
 import { IntlShape } from "react-intl";
 
 export const getLinkTypeOptions = (intl: IntlShape) =>
