+++ src/structures/components/MenuItemsSortableTree/MenuItemsSortableTree.tsx
@@ -1,5 +1,5 @@
 import { SortableTree } from "@dashboard/components/SortableTree";
-import { MenuTreeItem, RecursiveMenuItem } from "@dashboard/navigation/types";
+import { MenuTreeItem, RecursiveMenuItem } from "@dashboard/structures/types";
 import { UniqueIdentifier } from "@dnd-kit/core";
 import { Box, Text } from "@saleor/macaw-ui-next";
 import React from "react";
