+++ src/structures/components/MenuItemsSortableTree/MenuItemsSortableTreeItem.tsx
@@ -1,7 +1,7 @@
 import { TreeItemComponentProps } from "@dashboard/components/SortableTree/types";
 import { buttonMessages } from "@dashboard/intl";
-import { MenuItemType } from "@dashboard/navigation/components/MenuItemDialog";
-import { RecursiveMenuItem } from "@dashboard/navigation/types";
+import { MenuItemType } from "@dashboard/structures/components/MenuItemDialog";
+import { RecursiveMenuItem } from "@dashboard/structures/types";
 import { UniqueIdentifier } from "@dnd-kit/core";
 import { Box, Button, EditIcon, GripIcon, Text, TrashBinIcon } from "@saleor/macaw-ui-next";
 import React from "react";
