+++ src/structures/components/MenuListPage/MenuListPage.tsx
@@ -3,7 +3,7 @@ import { ListPageLayout } from "@dashboard/components/Layouts";
 import { MenuFragment } from "@dashboard/graphql";
 import useNavigator from "@dashboard/hooks/useNavigator";
 import { sectionNames } from "@dashboard/intl";
-import { menuListUrl, MenuListUrlSortField } from "@dashboard/navigation/urls";
+import { menuListUrl, MenuListUrlSortField } from "@dashboard/structures/urls";
 import { ListActions, PageListProps, SortPage } from "@dashboard/types";
 import { Box, Button, ChevronRightIcon, Text } from "@saleor/macaw-ui-next";
 import React from "react";
