+++ src/structures/components/MenuProperties/index.ts
