+++ src/structures/types.ts
