+++ src/structures/urls.ts
