+++ src/structures/views/MenuDetails/index.tsx
