+++ src/structures/views/MenuDetails/successHandlers.ts
