+++ src/structures/views/MenuDetails/utils.ts
