+++ src/structures/views/MenuList/sort.ts
@@ -1,6 +1,6 @@
 // @ts-strict-ignore
 import { MenuSortField } from "@dashboard/graphql";
-import { MenuListUrlSortField } from "@dashboard/navigation/urls";
+import { MenuListUrlSortField } from "@dashboard/structures/urls";
 import { createGetSortQueryVariables } from "@dashboard/utils/sort";
 
 export function getSortQueryField(sort: MenuListUrlSortField): MenuSortField {
